pyproject.toml
[project]
name = "better-face-tracking"
version = "0.1.0"
description = "Better face tracking and follow-speaker control for Canon PTZ cameras"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.4.0",
    "mediapipe>=0.10.0",
    "opencv-python-headless>=4.8.0",
    "requests>=2.31.0",
    "numpy>=1.24.0",
    "ndi-python>=0.1.0",
]

[project.scripts]
better-face-tracking = "main:main"

[tool.pyinstaller]
# Referenced by better-face-tracking.spec
name = "BetterFaceTracking"
bundle_id = "com.bft.better-face-tracking"
